storage-blob-delete: Support multiple parameters to d

+++ az-storage-blob-delete.sh
@@ -9,23 +9,30 @@ require STORAGE_ACCOUNT_NAME
 require STORAGE_ACCOUNT_CONTAINER
 require STORAGE_ACCOUNT_KEY
 
-set +e
-output=$(/usr/local/bin/az storage blob delete \
-	--account-key ${STORAGE_ACCOUNT_KEY} \
-	--account-name ${STORAGE_ACCOUNT_NAME} \
-	--container-name ${STORAGE_ACCOUNT_CONTAINER} \
-	--name ${1} 2>&1)
-status=$?
-set -e
-
-if [ $status -ne 0 ]; then
-	case "$output" in
-	*ErrorCode:BlobNotFound*)
-		echo "Blob not found, continuing..."
-		;;
-	*)
-		echo "An error occurred: $output"
-		exit $status
-		;;
-    esac
+if [ $# -eq 0 ]; then
+	echo "Usage: $0 blob_name [blob_name ...]" >&2
+	exit 1
 fi
+
+for blob_name in "$@"; do
+	set +e
+	output=$(/usr/local/bin/az storage blob delete \
+		--account-key "${STORAGE_ACCOUNT_KEY}" \
+		--account-name "${STORAGE_ACCOUNT_NAME}" \
+		--container-name "${STORAGE_ACCOUNT_CONTAINER}" \
+		--name "${blob_name}" 2>&1)
+	status=$?
+	set -e
+
+	if [ $status -ne 0 ]; then
+		case "$output" in
+		*ErrorCode:BlobNotFound*)
+			echo "Blob '${blob_name}' not found, continuing..."
+			;;
+		*)
+			echo "An error occurred deleting '${blob_name}': $output"
+			exit $status
+			;;
+	    esac
+	fi
+done
